Build options page with new-user logic
Now when you try to post the first time, the options page will open and allow
you to connect any services you like.

The OAuth flow continues to be a little wonky, and it may be simpler to rebuild
the client-side flow to handle authing more than once service. The OAuth2 lib
assume you only need to auth once and won't ever ask for more permissions.